+++ apps/server/src/invoice/invoice.controller.ts
@@ -251,8 +251,11 @@ export class InvoiceController {
 					details: error,
 				});
 			} else {
-				this.logger.error(`unknown error: ${error.message}\n${error.stack}`);
-
+				if (error instanceof Error) {
+					this.logger.error(`unknown error: ${error.message}\n${error.stack}`);
+				} else {
+					this.logger.error('unknown error: ', error);
+				}
 				throw new InternalServerErrorException();
 			}
 		}